move update part to own component
