+++ src/components/settings.tsx
@@ -1,9 +1,6 @@
 import { Description } from '@radix-ui/react-dialog';
 import { useEffect, useId, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { updaterConfig } from '@/config/updater';
-import type { UpdateCheckResult } from '@/lib/update-checker';
-import { checkForUpdate, normalizeVersion } from '@/lib/update-checker';
 import {
     getActiveProjectFromState,
     useCaaSConfigStore,
@@ -11,10 +8,10 @@ import {
 import { useSettingsStore } from '@/stores/settings-store';
 import type { ModeType } from '@/types/mode';
 import type { StageType } from '@/types/stage';
-import { version } from '../../package.json';
 import Icon from './icons/icon';
 import ManageProjectsDialog from './manage-projects-dialog';
 import { ProjectSettingsSection } from './project-settings-section';
+import { UpdateSection } from './update-section';
 import { Button } from './ui/button';
 import { CheckboxFieldWithTooltip } from './ui/checkbox-field-with-tooltip';
 import {
@@ -62,9 +59,6 @@ function Settings() {
 
     const [isSettingsOpen, setIsSettingsOpen] = useState(false);
     const [isManageProjectsOpen, setIsManageProjectsOpen] = useState(false);
-    const [isCheckingUpdate, setIsCheckingUpdate] = useState(false);
-    const [updateCheckResult, setUpdateCheckResult] =
-        useState<UpdateCheckResult | null>(null);
 
     // Local state for deferred save
     const [localCustomerName, setLocalCustomerName] = useState(
@@ -83,7 +77,7 @@ function Settings() {
     const [localCount, setLocalCount] = useState(count);
 
     // Sync local state from store when dialog opens
-    // biome-ignore lint/correctness/useExhaustiveDependencies: We just need to set the active selection when the settings are open
+    // biome-ignore lint/correctness/useExhaustiveDependencies: intentionally only sync on open
     useEffect(() => {
         if (isSettingsOpen) {
             setLocalCustomerName(activeSelection?.customerName ?? '');
@@ -138,10 +132,6 @@ function Settings() {
         setIsSettingsOpen(false);
     };
 
-    const handleCancel = () => {
-        setIsSettingsOpen(false);
-    };
-
     const onAddProjectClick = () => {
         setIsSettingsOpen(false);
         setIsManageProjectsOpen(false);
@@ -150,23 +140,6 @@ function Settings() {
     const openManageProjectsDialog = () => {
         setIsSettingsOpen(false);
         setIsManageProjectsOpen(true);
-    };
-
-    const handleCheckForUpdate = async () => {
-        setIsCheckingUpdate(true);
-        setUpdateCheckResult(null);
-        try {
-            const result = await checkForUpdate(
-                version,
-                updaterConfig.latestReleaseApiUrl,
-                updaterConfig.releasesPageUrl
-            );
-            setUpdateCheckResult(result);
-        } catch {
-            setUpdateCheckResult({ status: 'error' });
-        } finally {
-            setIsCheckingUpdate(false);
-        }
     };
 
     return (
@@ -258,88 +231,14 @@ function Settings() {
 
                         <hr className="border-t my-4" />
 
-                        {/* Updates section */}
-                        <div className="flex flex-col gap-3 my-4">
-                            <div className="flex items-center justify-between">
-                                <div>
-                                    <h2 className="font-semibold">
-                                        {t('app.settings.dialog.update.label')}
-                                    </h2>
-                                    <span className="text-sm text-muted-foreground">
-                                        {t(
-                                            'app.settings.dialog.update.currentVersion',
-                                            { version }
-                                        )}
-                                    </span>
-                                </div>
-                                <Button
-                                    type="button"
-                                    variant="outline"
-                                    onClick={handleCheckForUpdate}
-                                    disabled={isCheckingUpdate}
-                                >
-                                    <Icon
-                                        icon="arrow-clockwise"
-                                        className="size-4"
-                                    />
-                                    {isCheckingUpdate
-                                        ? t(
-                                              'app.settings.dialog.update.checking'
-                                          )
-                                        : t(
-                                              'app.settings.dialog.update.checkButton'
-                                          )}
-                                </Button>
-                            </div>
-                            {updateCheckResult?.status === 'upToDate' ? (
-                                <p className="text-sm text-green-600">
-                                    {t('app.settings.dialog.update.upToDate', {
-                                        currentVersion:
-                                            normalizeVersion(version),
-                                        latestVersion:
-                                            updateCheckResult.latestVersion,
-                                    })}
-                                </p>
-                            ) : null}
-                            {updateCheckResult?.status === 'updateAvailable' ? (
-                                <div className="text-sm text-yellow-600 flex items-center gap-2">
-                                    <span>
-                                        {t(
-                                            'app.settings.dialog.update.updateAvailable',
-                                            {
-                                                currentVersion:
-                                                    normalizeVersion(version),
-                                                latestVersion:
-                                                    updateCheckResult.latestVersion,
-                                            }
-                                        )}
-                                    </span>
-                                    <a
-                                        href={updateCheckResult.releaseUrl}
-                                        target="_blank"
-                                        rel="noreferrer"
-                                        className="underline"
-                                    >
-                                        {t(
-                                            'app.settings.dialog.update.viewRelease'
-                                        )}
-                                    </a>
-                                </div>
-                            ) : null}
-                            {updateCheckResult?.status === 'error' ? (
-                                <p className="text-sm text-red-600">
-                                    {t(
-                                        'app.settings.dialog.update.checkFailed'
-                                    )}
-                                </p>
-                            ) : null}
-                        </div>
+                        <UpdateSection />
+
                         <hr className="border-t my-4" />
                         <DialogFooter>
                             <Button
                                 type="button"
                                 variant="outline"
-                                onClick={handleCancel}
+                                onClick={() => setIsSettingsOpen(false)}
                             >
                                 {t('app.settings.dialog.cancel')}
                             </Button>
